+++ Changelog.rst
@@ -5,10 +5,14 @@ version 3.15.1
 
 * Performance improvements to `cf.aggregate`
   (https://github.com/NCAS-CMS/cf-python/issues/640)
+* New methods: `cf.Data.argmin`, `cf.Field.argmin`
+  (https://github.com/NCAS-CMS/cf-python/issues/577)
+* Fix bug when using the ``-d`` option to the `cfa` script 
+  (https://github.com/NCAS-CMS/cf-python/issues/649)
 * Changed dependency: ``cfunits>=3.3.6``
 
- ----
-  
+----
+
 version 3.15.0
 --------------